added in package private for query and modelview classes with same logic.
