+++ dbflow-processor/src/main/java/com/raizlabs/android/dbflow/processor/definition/ModelViewDefinition.java
@@ -15,7 +15,9 @@ import com.raizlabs.android.dbflow.processor.definition.method.MethodDefinition;
 import com.raizlabs.android.dbflow.processor.definition.method.PrimaryConditionMethod;
 import com.raizlabs.android.dbflow.processor.handler.DatabaseHandler;
 import com.raizlabs.android.dbflow.processor.model.ProcessorManager;
+import com.raizlabs.android.dbflow.processor.utils.ElementUtility;
 import com.raizlabs.android.dbflow.processor.utils.StringUtils;
+import com.raizlabs.android.dbflow.processor.validator.ColumnValidator;
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.FieldSpec;
@@ -121,11 +123,22 @@ public class ModelViewDefinition extends BaseTableDefinition {
 
     @Override
     protected void createColumnDefinitions(TypeElement typeElement) {
-        List<? extends Element> variableElements = manager.getElements().getAllMembers(typeElement);
+        List<? extends Element> variableElements = ElementUtility.getAllElements(typeElement, manager);
+        ColumnValidator columnValidator = new ColumnValidator();
         for (Element variableElement : variableElements) {
             if (variableElement.getAnnotation(Column.class) != null) {
-                ColumnDefinition columnDefinition = new ColumnDefinition(manager, variableElement, this, false);
-                columnDefinitions.add(columnDefinition);
+                // package private, will generate helper
+                boolean isPackagePrivate = ElementUtility.isPackagePrivate(element);
+                boolean isPackagePrivateNotInSamePackage = isPackagePrivate && !ElementUtility.isInSamePackage(manager, element, this.element);
+
+                ColumnDefinition columnDefinition = new ColumnDefinition(manager, variableElement, this, isPackagePrivateNotInSamePackage);
+                if (columnValidator.validate(manager, columnDefinition)) {
+                    columnDefinitions.add(columnDefinition);
+
+                    if (isPackagePrivate) {
+                        columnDefinitions.add(columnDefinition);
+                    }
+                }
 
                 if (columnDefinition.isPrimaryKey || columnDefinition instanceof ForeignKeyColumnDefinition || columnDefinition.isPrimaryKeyAutoIncrement) {
                     manager.logError("ModelViews cannot have primary or foreign keys");
